requirements.txt
watchdog>=4.0.0
python-frontmatter>=1.1.0
requests>=2.31.0
python-dotenv>=1.0.0
filelock>=3.13.0
langfuse>=2.0.0
neo4j>=5.0.0
# API (Phase 6)
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
httpx>=0.27.0
python-multipart>=0.0.9
pydantic>=2.0.0
# Discord Bot (Phase 8)
discord.py>=2.0.0
# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
anyio[trio]>=4.0.0
